tmdb2/requirements.txt
asyncio== 3.4.3
aiohttp==3.10.10
streamlit==1.39.0
openai==1.51.2
python-dotenv==1.0.1
pandas==2.2.3
scikit-learn==1.5.2
fuzzywuzzy==0.18.0
requests==2.32.3
streamlit-option-menu==0.4.0
python-Levenshtein==0.12.0  # Required for fuzzywuzzy
